The runnableSpectrumLite: contains the current version that can be copied into your computer and should be able to run. Default look is switched to Nimbus. You can try downloading & using SpectrumLite now.
Database management Utility has been shaped with basic functions:

1. Refresh: data tree collapses to show only databases. Any change of
the databases inside or outside of SpectrumLite will be automatically
recognize

2. New database: create an empty database with your choice of name. Name
extension not ends with ".db" will be automatically changed ".db".
Database Name has to follow the standard of windows files: characters
including \ / : * ? " < > | are not allowed.

3. Import databases: can import single or multiple databases (.db files)

4. Import Tables: can import single or multiple files into SpectrumLite
as tables. ".csv" and FVS file ".yld" are automatically recognized. With
other files with extensions SpectrumLite currently does not recognize,
it will ask for users to define the ".something" would be TAB, SPACE, or
COMMA delimited. ASK is one time and the format will be applied to all
files with the same extension.

5. Rename: database or table: tables in our system so we ALLOW special
characters in the name, although naming table with special characters is
not welcomed.

6. Copy: Within SpectrumLite software, current tables and databases can
be selected for copied and pasted into another existing database. Users
can select single or multiple items to be copied. 

7. Delete:  Users can select single or multiple items to be deleted. 


Potential Future update:
1. Copy and import speed is slow: 5sec/yield file. So next update may
need "INSERT BATCH"

2. Currently not support create table and edit table. Need to add these
functions

3. A progress bar to notify users should be needed. Currently when
import or copy, the process is slow and users dont know when the task
finished so they can continue with using the software

Current task: we should to switch our focus on Data analysis,
Formulation, and GUi for the main optimization function now. Other
improvements on Database Management Utility will be after that. 



  



Former-commit-id: 4d5048ceacfbd7656df0976215ec42a7e4868732
Former-commit-id: 39a9760394a63f76065e6e23f097e925cb552509